# src/config/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Api documentation for GooSendr. A logistics application.<br><br><br> <b>NEW</b><br> 1. Removal of username field <br> 2. Addition of the courier api<br><br>",
    "version": "1.0.0",
    "title": "GooSendr Api Documentation",
    "termsOfService": "",
    "customeSiteTitle": "GooSendr Api Docs",
    "contact": {
      "email": "aditech5170@gmail.com"
    }
  },
  "host": "52.14.168.100:3000",
  "basePath": "/api/v1",
...
  "schemes": [
    "http"
  ],
  "paths": {
    "/user/register": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Create user",
        "description": "Registering a new user.",
        "operationId": "createUser",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Created user object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/registration.model"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/auth.model"
            }
          },
          "400": {
            "description": ""
          }
        }
      }
    },
...
    "/user/forgot-password": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Forgot password",
        "description": "",
        "operationId": "forgotPassword",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "The user email registered in the system. ",
            "required": true,
            "schema": {
              "$ref": "#/definitions/forgotPassword.model"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "UNAUTHORIZED"
          },
          "401": {
            "description": "UNAUTHORIZED"
          }
        }
      }
    },
    "/user/reset-password": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Reset password with token from user's email",
        "description": "",
        "operationId": "resetPassword",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Token and Password",
            "required": true,
            "schema": {
              "$ref": "#/definitions/resetPassword.model"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": ""
          },
          "401": {
            "description": "UNAUTHORIZED"
          }
        }
      }
    },
    "/user/change-password": {
      "patch": {
        "tags": [
          "user"
        ],
        "summary": "Changing of password",
        "description": "This can only be done by the logged in user.",
        "operationId": "changePassword",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/changePassword.model"
            }
          }
        ],
        "security": [
          {
            "userToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "invalid old password"
          },
          "401": {
            "description": "UNAUTHORIZED"
          }
        }
      }
    },
    "/user/update-user": {
      "patch": {
        "tags": [
          "user"
        ],
        "summary": "Update user Information",
        "description": "This can only be done by the logged in user.",
        "operationId": "updateUser",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/updateInfo.model"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/info.model"
            }
          },
          "400": {
            "description": ""
          },
          "401": {
            "description": "UNAUTHORIZED"
          }
        },
        "security": [
          {
            "userToken": []
          }
        ]
      }
    },
...
    "/address/{id}": {
      "get": {
        "tags": [
          "user address book"
        ],
        "summary": "get user individual address",
        "description": "",
        "operationId": "getAddress",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "id of registered user",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/addressOutput.model"
            }
          },
          "400": {
            "description": "UNAUTHORIZED"
          }
        },
        "security": [
          {
            "userToken": []
          }
        ]
      },
      "patch": {
        "tags": [
          "user address book"
        ],
        "summary": "update individual address",
        "description": "",
        "operationId": "updateAddress",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "id of registered user",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/addressOutput.model"
            }
          },
          "400": {
            "description": "UNAUTHORIZED"
          }
        },
        "security": [
          {
            "userToken": []
          }
        ]
      },
      "delete": {
        "tags": [
          "user address book"
        ],
        "summary": "delete individual address",
        "description": "",
        "operationId": "delete_Address",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "id of registered user",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "UNAUTHORIZED"
          }
        },
        "security": [
          {
            "userToken": []
          }
        ]
      }
    }
  },
  "securityDefinitions": {
    "userToken": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    },
    "courierToken": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "definitions": {
    "registration.model": {
      "type": "object",
      "properties": {
        "fullName": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        },
        "countryCode": {
          "type": "string",
          "description": "eg +233"
        },
        "avatar": {
          "type": "object"
        }
      }
    },
...
    "updateInfo.model": {
      "type": "object",
      "properties": {
        "fullName": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        },
        "facebook": {
          "type": "object"
        },
        "google": {
          "type": "object"
        },
        "countryCode": {
          "type": "string",
          "description": "eg +233"
        },
        "avatar": {
          "type": "object"
        }
      }
    },
...
    "info.model": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string"
        },
        "fullName": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        },
        "facebook": {
          "type": "object"
        },
        "google": {
          "type": "object"
        },
        "avatar": {
          "type": "object"
        }
      }
    },
...
    "auth.model": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string"
        },
        "fullName": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        },
        "facebook": {
          "type": "object"
        },
        "google": {
          "type": "object"
        },
        "avatar": {
          "type": "object"
        },
        "token": {
          "type": "string"
        }
      }
    },
...
    "resetPassword.model": {
      "type": "object",
      "properties": {
        "token": {
          "type": "string"
        },
        "newPassword": {
          "type": "string"
        },
        "verifyPassword": {
          "type": "string"
        }
      }
    },
    "changePassword.model": {
      "type": "object",
      "properties": {
        "newPassword": {
          "type": "string"
        },
        "oldPassword": {
          "type": "string"
        }
      }
    },
...
    "forgotPassword.model": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        }
      }
    },
...
    "addressOutput.model": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string"
        },
        "fullName": {
          "type": "string"
        },
        "address1": {
          "type": "string"
        },
        "address2": {
          "type": "string"
        },
        "district": {
          "type": "string"
        },
        "region": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        },
        "countryCode": {
          "type": "string"
        },
        "phone2": {
          "type": "string"
        },
        "countryCode2": {
          "type": "string"
        },
        "createdAt": {
          "type": "string"
        },
        "user": {
          "type": "object"
        }
      }
    },
...
  }
}